Remove 'UserDefinedConstraint'
UserDefinedConstraint was a stub for future functionality, but is
interfering with refactoring. It will likely return when we have a
concrete use case to guide its implementation.

Removing UserDefinedConstraint also means that DefinedConstraint and
SystemDefinedConstraint are redundant, so this change also removes
DefinedConstraint and replaces it with SystemDefinedConstraint (reducing
the level of OR constraints on the query)